Offer an option to append or prepend the line
Summary:
The output is clearly more legible when the container name is put
before the text, especially for ps where arguments could be long.

By default, the command prepends 20 characters with container name
and spaces. This behavior can be modified by the `--append` flag
and the `-w <number of characters>` argument.

Ref T1456

Test Plan:
- test prepend/append modes
- check how container name is truncated
- non UTF-8 console (… is replaced by +)

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1456

Differential Revision: https://devcentral.nasqueron.org/D1837